moved cbd-article to common folder
