+++ app/components/common/cbd-article.vue
