+++ app/views/portals/forum-id.vue
@@ -127,7 +127,7 @@
 import ForumsApi from '~/api/forums';
 import ArticlesApi from '~/api/articles';
 import jumpToAnchor from '~/services/jump-to-anchor.js';
-import  cbdArticle  from './cbd-article.vue';
+import  cbdArticle  from '../../components/common/cbd-article.vue';
 import Post from '~/components/forums/post.vue';
 import EditPost from '~/components/forums/edit-post.vue';
 import pending   from '~/services/pending-call'
